Do not process gcloudignore files when uploading support-packages
This is a minor change, we are getting some error message on the upload
support packages steps from the CPaaS workflows because the
google-github-actions/upload-cloud-storage action expects a gcloudignore
to be present by default. Since we don't expect to process this kind of
file, we just set the option to false on all occurrences of the action.

Without this change the actions still worked correctly, this is merely
to get rid of the annoying error messages in the CI output.