Fix Konflux build after collector scripts removal
PR #1814 removed the scripts used by collector to run, making it
possible for it to be run directly. The Konflux build is still
attempting to copy the scripts directory into the image even though it
doesn't exist anymore, so we should remove that line from the
dockerfile.

As an additional cleanup, since we have been linking falco statically
for some time now, there's no need to call ldconfig anymore.