Add collector version to compilation step
Yet another mistake from #1814. The version of collector is now
hardcoded into the binary, but that needs to be set at compile time,
which was not being done in konflux builds. The label with the version
was being set properly, this only affects collector printing its version
to the logs.