Build the collector binary directly in the docker image build
Historically, we've always built the collector binary inside a running
container and then copied it from the host to the final image. This
change makes it so the builder image is used as a first step to compile
the collector binary, then the binary is copied over to the final image
in a second stage.

Development workflow should not change too much, make image will still
compile collector, run the unit tests and generate a final image.
Multiple successive compilations are sped up by using a cache mount for
the build directory.

Main benefit of the change is getting rid of the make targets that run
cmake while exec'ed into the builder image, these don't really make much
sense and are quite a bit more complex than they need to be IMO. Of
note, execing into a builder image and running the cmake commands
directly in there will still work, if that is the workflow of choice.

Second benefit is making image builds a bit closer to what konflux and
downstream do, but not overly so, maybe one day we'll be able to have
all image builds come from a single source of truth.