Replace processes-listening-on-ports.c with plop.py
Recently stumbled onto a flake of the TestProcessListeningOnPort test
and decided to look into it a bit further. I noticed the way works is by
using a file that is read by a C binary, once the file is read, the
binary deletes it. I believe the flake is due to the file being created
via an exec into the container, then checking if the file exists on a
mounted directory, which could very well lead to a race condition with
the exec being done, but the file not showing up on the local file
system, leading to believe it has already been processed and it being
overwritten.

In order to have a more reliable way of opening and closing ports, the C
binary has been replaced with a flask app that can open ports on its
/open endpoint and closing them with its /close endpoint. These
operations should be a little more reliable, since the http.Get
operations will only return once the socket has been properly opened or
closed and we will get confirmation via the HTTP response code.

Some additional changes are needed for tests to pass with the new
version of the container, but most are minor.