Use archives for test logs
This is due to the ansible fetch module either flattening files directly
onto the destination directory, overwritten files with the same name, or
keeping a long path that includes the hostname and the full path to the
files. Best workaround I could come up with is to generate uniquely
named archives that hold all logs with the proper paths and files, then
move those with the flat option.