Unarchive log files before uploading artifacts
Doing this makes the artifacts easier to use, since it doesn't require
unzip and unarchiving the logs separately. It also fixes an error with
the benchmark steps, since these expect the perf.json files to be
readily accessible.

Unfortunately, I wasn't able to do this in ansible, so running the test
manually on remote systems still creates archives on the local host.