Output the number of times a log has been throttled (#1876)
Adding the number of times a log message has been throttled can add
value in the sense that having the message show up once or twice may not
be as worrisome as it showing up thousands of times.

In order to achieve this, the log message can be split between its
header and the actual message. The header will now be handled by classes
separate to LogMessage that will hold the information necessary to print
them and use a common ILogHeader interface. This is done in order to
reduce the amount of code necessary in macros, there shouldn't be a huge
performance difference, but the current approach might take a bit more
memory in order to store this data, some testing is needed.

Co-authored-by: Olivier Valentin <ovalenti@redhat.com>