Remove --platform from start-builder target (#1922)
Podman has a bug in it which makes it so when using the --platform
argument, it will ignore existing images in cache and always attempt to
pull. This is particularly annoying when trying to test a change to the
builder image, since running the start-builder target will:
- Build the image from source.
- Pull the image from quay.
- Run the image without the local modifications.

Removing the argument fixes the issue and it should have no major impact
on developer workflow, since it is meant to be used locally anyways.