Prevent pulling builder when building from source
This is a workaround to a podman issue in which, adding --platform to a
pull command, will cause podman to always pull the image. This is
particularly annoying when testing changes to the builder image, since
setting it to run will overwrite the local image and pull an existing
one instead.