Enable building with vcpkg
This enables building collector outside of the builder container by
leveraging the [vcpkg](https://vcpkg.io) package manager. This is not
ideal for CI, since vcpkg pulls from the internet so isolated builds
are not possible, but it is more convenient than having to install all
libraries to the system.

A simple CMakePresets.json file was added to enable even easier use.
Command for using this build is: cmake --preset=vcpkg