Prevent overwriting test logs on multi VM tests
When running integration tests on multiple VMs with ansible, the logs
generated may be overwritten as the different VMs end. This also causes
the reports on the test summary to only show the result for the VMs that
finish last. This situation is fine if all VMs have the same failures,
but when only one of them is failing, (i.e: a verifier issue on a
specific kernel) the error can get masked and incorrectly show all tests
have passed.

In order to fix this issue, logs and reports generated by CI should now
have the VM name as part of their file name.