Build the builder image on native systems
In order to cut down on the time taken to build the builder image, which
is currently very long due to QEMU being used on small GHA VMs, we can
move the non x86 images to be built on remote VMs hosted on GCP and IBM
cloud. We already use these VMs for integration tests and for building
s390x images, this is just extending the functionality for ppc64le and
arm64.