Cap parallelism of dependency building
When building the third party dependencies, we use the nproc command to
try and get as many parallel builds as possible running at once. This
can bring problems on systems with large numbers of cores and not enough
ram to come with them, so for those we provide the option of manually
setting the number of parallel builds with an environment variable.