Move deprecated versions to RELEASED_VERSIONS.unsupported
The 3.74 and 4.3 versions of StackRox have been EOL for some time now
and we haven't built support packages for them since October and August
of last year respectively. These lines need to be moved because
otherwise the support package download page still lists the old packages
that are now outdated and missing newer kernels.