Add note about GRPC_TRACE
We've struggled in the past with connectivity issues and lack of
information from the GRPC library, e.g. it's hard to distinguish a
timeout error from a certificate error -- Collector will just say "can't
connect" in both cases.

Add a note about useful workaround for that with GRPC_TRACE environment
variable into the troubleshooting section.