Pin shfmt version (#2055)
The lint CI job has started failing because the go version required by
the latest release of shfmt is higher than the one being installed on
the fedora:40 image. We could bump the image, but for now I think
pinning the shfmt version would lead to the least pain in the future.