Fix QA tag missing collector version in konflux tests (#2057)
The QA tag being used for konflux tests on PRs is malformed, leading to weird errors. This should fix it.

When running on a PR that bumps the QA_TAG, the test containers include the collector tag as a suffix in order to prevent collisions with other PRs that might be bumping the version, so they would look something like this: 2.0.3-3.21.x-26-g5a61d712ab-fast

However, on konflux, the collector tag was not being properly set and the tags looked like this: 2.0.3-