Allow running integration tests with a CRI client (#2048)
* Allow running integration tests with a CRI client

Add a new CRI executor which allows controlling containers through a
standard CRI interface. This allows us to run our integration tests
using the containerd (should also allow CRI-O, but it is currently
untested).

The main reason for this change is that Garden Linux has stopped
shipping docker in its VMs and made them immutable, so the only way to
properly test Garden kernels is to either use containerd or create our
own image of it, we clearly chose the first of the two options.

* Bump qemu version

* Minor cleanups

* More cleanups

- Make CNI and crictl versions easier to change.
- Add comment explaining the source of CNI configuration file.
- Better error message if we fail to read the body of a plop response