Fix CPU profiler
When we removed the /module mount from the collector daemonset we
inadvertedly broke the gperftool cpu profiler, since it was using that
directory for dumping its output files. This patch moves the dump
directory to /var/profiles, which will be a new emptydir mount on the
collector daemonset.

A new test is also added to ensure the profiles still starts, stops and
produces some output.