Stop running the unit tests as part of the image build
When iterating on collector changes, waiting for the image build takes
some time because we build the binary, then run the unit tests which
also attempt to build the binary again and run the tests themselves. If
we ask for the image to be built, we should only do the steps required
to build the collector binary and creating the image, if we want to run
unit tests, we have a separate target for it.

On CI, the unit tests are already being run as a separate step, so
removing it from the image build should be a net win with no losses
there.

We are also removing the unit tests from the konflux build, since they
are redundant with the upstream CI job. The images produced by konflux
are verified with integration tests.