Stop making copies of hostname
The hostname for a given node is stored in the HostInfo singleton after
we first try to access it and will outlast any other object attempting
to use it. As such, we should try to avoid unnecessary copies and just
call the GetHostname method of the HostInfo class.

The GetHostname function from out utilities has been turned into a
static method of HostInfo so we can directly access it calling
HostInfo::GetHostname(), otherwise we would need
HostInfo::Instance().GetHostname() which is a bit too verbose.