Replace all include guards with #pragma once
Since we don't have a common nomenclature for include guards, whenever a
new header file is added to the project, it is up to the writer to come
up with the name for the guard. This becomes a bit annoying when trying
to figure out what the name should be and comes off a bit sloppy, since
different files follow different guidelines. Using #pragma once solves
both these problems.

The only header file that hasn't been changed is optionparser.h, since
that is actually a third party header file that should not be modified
manually.