Refactor NetworkStatusNotifier (#2089)
* Refactor NetworkStatusNotifier

This change started out from the effort to unify the services used to
communicate with sensor. While attempting to add an output object to the
notifier, it became apparent this class has some issues with ownership
of its members, which in turn makes it a bit odd to extend and modify.

In order to alleviate the problem the following changes are introduced:
- The connection scraper is now owned by NetworkStatusNotifier.
- The communication channel for network info is now owned by
  NetworkStatusNotifier.
- Connection stats are now owned by NetworkStatusNotifier.
- Move ownership of the prometheus registry into CollectorStatsExporter.
- Refactor CollectorConnectionStats into its own file.

When stating "X is now owned by Y" it stands for X now is part of Y and
its lifecycle is fully handled by Y (i.e X is created with Y and
destroyed with Y, preferably following RAII).

* Access configuration variables from configuration directly

NetworkStatusNotifier keeps a reference to our configuration object, but
also makes copies of several trivial variables. Accessing them directly
from the configuration reference should be preferred for consistency and
it would make it easier in the future if we want to implement some sort
of hot reloading for configuration.