Add hotreload.sh
This script is inspired by the hotreload mode in the main repo.

When using a k8s cluster with access to the local filesystem (like kind,
k3s, minikube, etc.), it configures the collector container in the
collector daemonset to execute a locally built binary, which is useful
for iterating fast or running with sanitizers enabled.

The script also provides a --revert option for dropping the
configuration changes once they are no longer needed.