Make IsCanonicalExternalIp independent of system endianness (#2113)
IsCanonicalExternalIp does not work on big endian systems like s390x.
This change makes it so the check is always done in network endianness,
effectively making it independent of the system. Change is only needed
on IPv4 checks because IPv6 uses 64 bit integers filled with 1s, so
endianness is irrelevant in that case.

Co-authored-by: Jouko Virtanen <jvirtane@redhat.com>