Add a release workflow (#2125)
This new workflow will enable people working in the collector repo to
simply create a new release by issuing a workflow from the github
actions tab. Two inputs can be provided to the workflow:
- version: Specify the <Major>.<minor> version to be tagged.
- dry-run: Run the workflow, create all tags but don't push to origin.

The intended usage is to make a first run with the dry-run flag on to
check everything works as expected, then make a second run without the
dry-run flag to create the release tags.

If the version input is left unchanged a new minor version will be
tagged, creating the <Major>.<minor>.x tag on the master branch, setting
up the release-<Major>.<minor> branch with an empty commit and the new
.0 tag on top of said commit.

If a version is provided the behavior depends on what the actual value
is. The workflow will check if the behavior intended by the caller is to
do a major, minor or patch release by checking the provided version
against the latest tag on master. In all cases, the workflow ensures the
release will always have the correct version numbers.