Build arm64 test container by default
After the changes introduced in #2106, we are building arm64 images by
default and trying to test them. If a change is made to the integration
tests, we are only building the amd64 version of the container and the
arm64 tests fail when trying to pull the image for that architecture.

This patch makes it so the test container is built for amd64 and arm64
by default.