Add hotreload.sh (#2110)
* Add hotreload.sh

This script is inspired by the hotreload mode in the main repo.

When using a k8s cluster with access to the local filesystem (like kind,
k3s, minikube, etc.), it configures the collector container in the
collector daemonset to execute a locally built binary, which is useful
for iterating fast or running with sanitizers enabled.

The script also provides a --revert option for dropping the
configuration changes once they are no longer needed.

* Add THREAD_SANITIZER build configuration

This new option will compile the collector binary with thread sanitizer
support. libtsan is added to the image-dev in order for collector to be
able to use it at runtime. libubsan is also added since it was missing.

* Add some docs

* Fix inconsistent flag handle and check collector is deployed