Add a label forcing a test container rebuild on PRs (#2180) (#2186)
Usually, the test container only needs rebuilding when a change to the
tests or related files happen. However, when creating PRs that target
old branches (like when backporting changes to a release branch), the
rebuild may not trigger and an older image may be subject to newer tests
it is not meant to pass.

In order to work around this, a label for triggering rebuilds of the
test container manually is added, allowing PRs to rebuild the test
container on demand.