Use go cross-compilation to speed up test builds on CI (#2210)
After adding Arm runners on GHA, we made it so the Arm test image is
always built. Unfortunately, because we use QEMU to build the images,
the build has gotten quite slow.

In order to speed up the build, this change makes it so we cross-compile
the binaries for all our supported platforms locally and the image build
simply copies the binaries into the final image.