Fetch VM families from GCP for testing
Few times over the last couple of years we have found ourselves having
to update the list of VMs we use for testing manually when a family is
added or removed from GCP. This is an annoying process that also turns
our CI red until we get it done when an image is removed.

With this change, the list of VM images to be used for RHEL and Ubuntu
will be directly retrieved from GCP with a script.