Backport P, Z and Arm runners to 3.21 (#2254)
* ROX-29116: (fix) Use ARM GH action workflow runners for ARM builds (#2106)

* Revert "Revert GH arm changes (#2107)"
* Add arm64  stable, beta, and dev channels for Google COS integration tests (#2096)

* Use P and Z GHA runners (#2215)

* Use go cross-compilation to speed up test builds on CI (#2210)

After adding Arm runners on GHA, we made it so the Arm test image is
always built. Unfortunately, because we use QEMU to build the images,
the build has gotten quite slow.

In order to speed up the build, this change makes it so we cross-compile
the binaries for all our supported platforms locally and the image build
simply copies the binaries into the final image.

---------

Co-authored-by: Robby Cochran <rcochran@redhat.com>