Always rebuild builder and test container on PRs to release branches
By nature, the release branches can drift from the master branch, this
includes differences to the dependencies installed to the builder image
and the tests the collector binary is run against.

Because the builder and integration tests images both take a pretty long
time to build, we originally settled on only rebuilding these when it
was absolutely necessary and added a couple PR labels to have them
rebuilt on demand. Since we now have Power and Z runners for faster
builds and the integration tests are being built in a faster manner too,
we should just rebuild these on PRs targeting the release branch to
reduce any potential confusion on colleagues that may not be aware of
this intricate configuration.