Remove SSH access to builder container
Z builds are currently failing when building the builder image with the
following error:

```
Dockerfile:72
--------------------
  71 |     # Add remote development user
  72 | >>> RUN useradd -m remoteuser \
  73 | >>>   && yes c0llectah | passwd remoteuser || if [[ $? -eq 141 ]]; then true; else exit $?; fi
  74 |
--------------------
ERROR: failed to build: failed to solve: process "/bin/sh -c useradd -m remoteuser   && yes c0llectah | passwd remoteuser || if [[ $? -eq 141 ]]; then true; else exit $?; fi" did not complete successfully: exit code: 1
gmake: *** [Makefile:27: builder] Error 1
```

In the past, SSH was used with CLion for development, but devcontainers
are now properly supported by that IDE, so there is no point in keeping
the SSH access method in the builder container.

The documentation section on developing with CLion has been removed
since it is no longer supported. I have not documented how to use a
devcontainer with that environment because I have no idea how that works
and don't have access to CLion for testing either, contributions to add
the section back in are welcomed.