Remove all driver related files (#2124)
* Remove RELEASED_VERSIONS and RELEASED_VERSIONS.unsupported

These files are used to correlate collector versions to stackrox
versions, main use for it was to create the index directory for
support-packages.

Since support-packages have been deprecated since 4.4, there's not much
point in keeping these files updated anymore and the only purpose it
seems to serve is taking up release engineers time ping-ponging a PR
that gets a rubber-stamped "approved" and merge from the collector team.
Hence, we are removing the files from the collector repo altogether.

Correlating versions between stackrox and collector can be achieved
using the following command in the main repo:
  `git show <tag>:COLLECTOR_VERSION`

I will add that command to our documentation as part of a larger
cleanup.

* Remove downstream driver workflows

And by extension all scripts, files and actions related to drivers.

* Remove CPaaS workflows completely