Fix cancel stats routine (#2677)
The stats go routine cancellation is implemented via callback for more
flexibility. But in case if the container was not yet in healthy state
and the stats go routine wasn't initialized yet, it might come to the
situation that we try to stop container without the callback. Verify
such use case.